% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_to_common_time.R
\name{align_to_common_time}
\alias{align_to_common_time}
\title{Align a dataframe to a common start date and time}
\usage{
align_to_common_time(
  dataset,
  subject_var,
  datetime_var,
  cutoff_datetime,
  timezone = "America/Chicago"
)
}
\arguments{
\item{dataset}{oftentimes this will be your TimeSeries or Circadian datatab}

\item{subject_var}{Animal or Subject ID}

\item{datetime_var}{Date created when importing your dataset}

\item{cutoff_datetime}{the first time you want included for this dataframe in "\%Y/\%m/\%d \%H:\%M:\%S" format}

\item{timezone}{Options other than America/Chicago are America/New_York, America_Los_Angeles, America/Chicago, Asia/Tokyo, Asia/Hong_Kong, Europe/London, Europe/Paris, and Europe/Berlin are popular options}
}
\value{
A new dataframe with all subjects starting at the same date/time
}
\description{
Align a dataframe to a common start date and time
}
\examples{
\dontrun{
data1_filt <- align_to_common_time(
  dataset = data1,
  subject_var = Animal,
  datetime_var = Date,
  cutoff_datetime = "2023/06/27 07:30:02",
  timezone = "America/Chicago"
)
}
}
