% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_to_photoperiod.R
\name{align_to_photoperiod}
\alias{align_to_photoperiod}
\title{Select photoperiod (light/dark cycle) for your time aligned dataframe}
\usage{
align_to_photoperiod(
  dataset,
  subject_var,
  datetime_var,
  lights_on = "07:00:00",
  lights_off = "19:00:00",
  cycle_length = 24,
  timezone = "America/Chicago"
)
}
\arguments{
\item{dataset}{your (combined) dataframe of timeseries or circadian data}

\item{subject_var}{Animal or Subject ID}

\item{datetime_var}{Date}

\item{lights_on}{time the lights turn on at}

\item{lights_off}{time the lights turn off at}

\item{cycle_length}{24 hours}

\item{timezone}{America/Chicago or others}
}
\value{
photoperiod aligned data
}
\description{
Select photoperiod (light/dark cycle) for your time aligned dataframe
}
\examples{
\dontrun{
# Using defaults
photoperiod_data1 <- align_to_photoperiod(dataset = data1, subject_var = Animal,
  datetime_var = Date)

# Overriding defaults
photoperiod_data1 <- align_to_photoperiod(dataset = data1, subject_var = Animal,
  datetime_var = Date, lights_on = "06:00:00", lights_off = "18:00:00")
  }
}
