% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_feeding_events.R
\name{calc_feeding_events}
\alias{calc_feeding_events}
\title{Calculate feeding events from the cumulative food in column}
\usage{
calc_feeding_events(
  df,
  id_col = "subject.id",
  time_col = "exp.hour",
  cumsum_col = "FoodInA_M"
)
}
\arguments{
\item{df}{dataframe}

\item{id_col}{subject.id}

\item{time_col}{exp.hour}

\item{cumsum_col}{FoodInA_M}
}
\value{
dataframe with new columns for rezeroed cumsum and each feeding event
}
\description{
This function rezeroes the data since we trimmed timepoints earlier.
Then it calculates each feeding event, with the amount eaten during the hour displayed.
}
\examples{
\dontrun{
df1 <- calc_feeding_events(df1)
}
}
