% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_PedMeters.R
\name{clean_PedMeters}
\alias{clean_PedMeters}
\title{Filter excessively large running events out}
\usage{
clean_PedMeters(
  df,
  id_col = "subject.id",
  time_col = "exp.hour",
  ped_col = "PedMeters_R",
  threshold = 50
)
}
\arguments{
\item{df}{dataframe}

\item{id_col}{subject.id}

\item{time_col}{exp.hour}

\item{ped_col}{PedMeters_R}

\item{threshold}{50 meters (or if you have wheel running, you can adapt this function)}
}
\value{
cleaned dataframe with NAs where excessively large events were
}
\description{
Filter excessively large running events out
}
\examples{
\dontrun{
df1_filtered <- clean_PedMeters(df1_filtered)
}

}
