% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_RER.R
\name{clean_RER}
\alias{clean_RER}
\title{Filter high and low RER values}
\usage{
clean_RER(df, rer_col = "RER_M", upper = 1.2, lower = 0.6)
}
\arguments{
\item{df}{dataframe}

\item{rer_col}{RER_M}

\item{upper}{1.2}

\item{lower}{0.6}
}
\value{
dataframe with cleaned RER data (NAs where values were outside params)
}
\description{
This is helpful to filter out torpor events. Relax thresholds if you know you're sending a mouse into ketogenesis
}
\examples{
\dontrun{
df1_filtered <- clean_RER(df1_filtered)
}

}
