% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_photoperiods.R
\name{create_photoperiods}
\alias{create_photoperiods}
\title{Create photoperiod columns for better graphing purposes}
\usage{
create_photoperiods(
  df,
  time_col = "exp.hour",
  period_length = 12,
  first_period = "Light"
)
}
\arguments{
\item{df}{dataframe}

\item{time_col}{exp.hour}

\item{period_length}{12}

\item{first_period}{Light}
}
\value{
dataframe with additional column for Light or Dark period
}
\description{
Create photoperiod columns for better graphing purposes
}
\examples{
df2 <- create_photoperiods(df1_filtered)


}
