% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_feeding_events.R
\name{filter_feeding_events}
\alias{filter_feeding_events}
\title{Title}
\usage{
filter_feeding_events(
  df,
  id_col = "subject.id",
  time_col = "exp.hour",
  cumsum_col = "FoodInA_M",
  threshold = 0.3
)
}
\arguments{
\item{df}{dataframe}

\item{id_col}{subject.id}

\item{time_col}{exp.hour}

\item{cumsum_col}{FoodInA_M}

\item{threshold}{0.3 (can be lower if you know how much they're really eating, but this is 300mg per event or more filtered out)}
}
\value{
cleaned dataframe with excessive feeding events filtered out and a new cumsum column
}
\description{
Title
}
\examples{
\dontrun{
df1_filtered <- filter_feeding_events(df1_filtered)
}


}
