% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_columns.R
\name{get_columns}
\alias{get_columns}
\title{Get all column names in a dataframe}
\usage{
get_columns(dataset, exclude = NULL)
}
\arguments{
\item{dataset}{combined dataframe}

\item{exclude}{optional columns to exclude can be entered here as a list =c("colum_name", ...)}
}
\value{
column names as a value
}
\description{
Get all column names in a dataframe
}
\examples{
\dontrun{
all_cols <- get_columns(df1)
}
}
