% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_boxplot.R
\name{graph_boxplot}
\alias{graph_boxplot}
\title{Create boxplots of summary data}
\usage{
graph_boxplot(
  dataset,
  group_var,
  columns_to_plot,
  subject_var,
  photoperiod_var = NULL,
  ylab = NULL,
  title = NULL,
  point_size = 3,
  point_alpha = 0.7
)
}
\arguments{
\item{dataset}{summary dataframe}

\item{group_var}{group or treatment}

\item{columns_to_plot}{column list or value of columns to use}

\item{subject_var}{subject.id}

\item{photoperiod_var}{NULL or photoperiod}

\item{ylab}{= NULL to just plot column title}

\item{title}{column title}

\item{point_size}{point size for each mean in the boxplot}

\item{point_alpha}{opacity of each point}
}
\description{
Once you create summary data, you can graph it all using boxplots.
You can choose to do this from photoperiod data or from general summary data.
This will allow you to compare light and dark cycles as well as parameters during each.
}
\examples{
plots <- graph_boxplot(dataset = summary_df2,group_var= group, columns_to_plot = cols, subject_var = subject.id)


}
