% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaboTYPE2.0-package.R
\docType{package}
\name{metaboTYPE2.0-package}
\alias{metaboTYPE2.0}
\alias{metaboTYPE2.0-package}
\title{metaboTYPE2.0: What the Package Does (One Line, Title Case)}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Meagan-Kingren/metaboTYPE2.0}
  \item \url{https://meagan-kingren.github.io/metaboTYPE2.0/}
  \item Report bugs at \url{https://github.com/Meagan-Kingren/metaboTYPE2.0/issues}
}

}
\author{
\strong{Maintainer}: Meagan Kingren \email{drkingren@gmail.com}

}
