% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ts.R
\name{plot_ts}
\alias{plot_ts}
\title{Basic time series graph}
\usage{
plot_ts(
  dataset,
  time_var,
  y_var,
  group_var,
  title = NULL,
  xlab = "Hour",
  ylab = NULL,
  shaded_days = NULL,
  shaded_width = 12,
  shaded_fill = "grey80",
  shaded_alpha = 0.5
)
}
\arguments{
\item{dataset}{your combined timeseries dataframe}

\item{time_var}{date}

\item{y_var}{column or parameter of choice}

\item{group_var}{animal or subject ID or treatment group}

\item{title}{"Time Series of (your parameter)}

\item{xlab}{hour}

\item{ylab}{column unit or null}

\item{shaded_days}{number of days to shade; select null for full experiment}

\item{shaded_width}{length of cycle}

\item{shaded_fill}{color of fill}

\item{shaded_alpha}{opacity of bars}
}
\value{
time series graph with dark cycle shaded
}
\description{
Basic time series graph
}
\examples{
\dontrun{
plot_ts(
  dataset = df1,
  time_var = exp.hour,
  y_var = kcal_hr_M,
  group_var = group,
  title = "Energy Expenditure",
  ylab = "EE (kcal/hr)"
)
}
}
