% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ts_all.R
\name{plot_ts_all}
\alias{plot_ts_all}
\title{Plot all numeric columns in a dataframe}
\usage{
plot_ts_all(
  dataset,
  time_var,
  group_var,
  y_cols = NULL,
  exclude = NULL,
  save_plots = FALSE,
  output_dir = "plots/",
  ...
)
}
\arguments{
\item{dataset}{timeseries data}

\item{time_var}{Date}

\item{group_var}{animal or subject ID or treatment group}

\item{y_cols}{leave null to autodetect numeric columns or specify from a list}

\item{exclude}{columns to skip in a dataframe}

\item{save_plots}{select TRUE to save plots}

\item{output_dir}{creates a new directory to save plots in}

\item{...}{Additional arguments passed to \code{plot_ts}}
}
\value{
graphs for every single column or every defined column
}
\description{
Plot all numeric columns in a dataframe
}
\examples{
\dontrun{
plots<- plot_ts_all(dataset = df1,
  time_var = exp.hour,
  group_var = group,
  save_plots = TRUE)
}
}
