% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_hello_user.R
\name{print_hello_user}
\alias{print_hello_user}
\title{Print hello user}
\usage{
print_hello_user(x)
}
\arguments{
\item{x}{This argument does nothing.
If it did do something, it might require more words to explain.}
}
\value{
Prints \code{"Hello, User"} to the console
}
\description{
This function just prints hello user.
}
\examples{
print_hello_user(x = 1)

}
