% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_subjects.R
\name{summarize_subjects}
\alias{summarize_subjects}
\title{Summarize each column in your dataframe}
\usage{
summarize_subjects(
  dataset,
  subject_var,
  columns_to_use,
  group_var,
  bodymass_var,
  photoperiod_var = NULL
)
}
\arguments{
\item{dataset}{dataframe}

\item{subject_var}{group or subject.id}

\item{columns_to_use}{all or select a specific list}

\item{group_var}{group}

\item{bodymass_var}{body.mass}

\item{photoperiod_var}{Null by default}
}
\description{
Summarize each column in your dataframe
}
\examples{
summary_df2_photo <- summarize_subjects( dataset = df2, subject_var =subject.id, columns_to_use = columns_updated, group_var = group, bodymass_var = body.mass, photoperiod_var = "photoperiod")



}
